+++ icu4c/source/test/cintltst/cstrcase.c
@@ -442,7 +442,7 @@ TestCaseFolding(void) {
 
     /* test simple case folding */
     p=simple;
-    for(i=0; i<sizeof(simple)/12; p+=3, ++i) {
+    for(i=0; i<(int32_t)sizeof(simple)/12; p+=3, ++i) {
         if(u_foldCase(p[0], U_FOLD_CASE_DEFAULT)!=p[1]) {
             log_err("error: u_foldCase(0x%04lx, default)=0x%04lx instead of 0x%04lx\n",
                     p[0], u_foldCase(p[0], U_FOLD_CASE_DEFAULT), p[1]);
